Minor mod to generate_command()
generate_command() now returns the full list of commands to execute.
After the commands are executed and the results generated, the second half of generate_command() will be callable in a new method().
This new method will be used to generate the empty files list and possibly move small files to a different location.