Modifications to _which() method
Modified _which() method to support unittests running on multiple platforms.
If _which() is passed the name of an executable, rather than a path, then the test to verify that the observed path matches the expected result is not performed.
This is because commands on the PATH such as ls and echo may be in different locations on other platforms.