Remove create directory for Data/Fastq
Previouly it was assumed that the user would supply the path to a run-directory, and the data would be located in the sub-directory 'Data/Fastq'.
Since we expect now to be passed the directory containing the files itself, and we can expect that it's a read-only directory, we don't need this code anymore.