Added check for RTAComplete.txt and RunInfo.xml
Added check for RTAComplete.txt and RunInfo.xml to Pipeline class. The idea is for Pipeline to perform this check as early as possible, so that an error can be passed to the user before Job objects are created.
Check is already present in Job objects, and is left included as a redundancy.