Added test for absense of required files.
Removed check for required files from ConvertJob.py. Pipeline.py will oversee requirements for all Jobs so that aborts can be raised earlier.
RTAComplete.txt and RunInfo.xml have been removed from the repository as they are now created and deleted on demand during testing.
Additional test added to test the creation of Pipeline objects when one of the two files is missing and verify an Error was raised.