Modified RTAComplete.txt and RunInfo.xml file-check (#65)
* Added check for RTAComplete.txt and RunInfo.xml

Added check for RTAComplete.txt and RunInfo.xml to Pipeline class. The idea is for Pipeline to perform this check as early as possible, so that an error can be passed to the user before Job objects are created.
Check is already present in Job objects, and is left included as a redundancy.

* Added test for absense of required files.

Removed check for required files from ConvertJob.py. Pipeline.py will oversee requirements for all Jobs so that aborts can be raised earlier.
RTAComplete.txt and RunInfo.xml have been removed from the repository as they are now created and deleted on demand during testing.
Additional test added to test the creation of Pipeline objects when one of the two files is missing and verify an Error was raised.

* Updated assertRaises to check for error message.

* Revert tests in test_creation()

Revert tests in test_creation() as the error strings need to be massaged before they can be compared.

* Added permissions test for RunInfo.xml

Added test to ensure RunInfo.xml is readable at the time it's detected.