WIP: Tracking sample processing failures in pipeline
Added code to Pipeline class to detect when a sample failed processing in the pipeline.
Based on the observation that most of the output files from each stage in processing are named after a sample id.
GenPrepFileJob auditing is not yet implemented.
Code is tested, but unit-tests are not yet written.
New test data should be created and deleted on demand, rather than remain static files.
Audit information will also be useful for plating multiple projects per plate.