Sample-sheets are now bound to Pipeline object.
A sample-sheet is now bound to the Pipeline object at init() time. This allows the sample-sheet to be validated once and stored, rather than processed for nearly every method. This reflects how Pipeline object is actually used in production; we don't reuse the Pipeline object to process multiple (run-directory, sample-sheet) tuples.
is_within_time_range() merged w/init() method, as it was the only method that called it, and in light of the above change.
validate(sample_sheet) method renamed to _validate_sample_sheet(), as it's now only called by init().
extract_metadata() method removed for being too generalized.
sample main removed as the best and most current example of how to use sequence_processing_pipeline is in qp-knight-lab-processing.
tests updated.
other minor changes as needed to support the above.