Detect when a sample failed processing in the pipeline. (#67)
* Detect when a sample failed processing in the pipeline.

Added code to help detect when a sample failed processing in the pipeline.
Based on the observation that most of the output files from each stage in processing are named after a sample id.
GenPrepFileJob auditing is not yet implemented.

* Migrated auditing of ConvertJob

* Auditing functionality migrated to Job classes

Auditing functionality migrated to Job classes from Pipeline class.
Unit-tests rewritten to support this action.

* Testing directory is now properly created and deleted.

* Sample-sheets are now bound to Pipeline object.

A sample-sheet is now bound to the Pipeline object at init() time. This allows the sample-sheet to be validated once and stored, rather than processed for nearly every method. This reflects how Pipeline object is actually used in production; we don't reuse the Pipeline object to process multiple (run-directory, sample-sheet) tuples.
is_within_time_range() merged w/init() method, as it was the only method that called it, and in light of the above change.
validate(sample_sheet) method renamed to _validate_sample_sheet(), as it's now only called by init().
extract_metadata() method removed for being too generalized.
sample main removed as the best and most current example of how to use sequence_processing_pipeline is in qp-knight-lab-processing.
tests updated.
other minor changes as needed to support the above.

* Removed SequenceDirectory class, test.

Removed SequenceDirectory class as it contains no functionality that is currently being used or hasn't been duplicated.
Tests for SequenceDirectory were removed as well.

* Modifications to support qp-klp

* flake8

* Pipeline.warnings is now properly initialized

* Adjusted substring matching in audit() methods

Adjusted substring matching in audit() methods based on real-world tests.

* flake8

* Removed timestamping, audit() now in base class

Removed timestamp-checking on run-directories. There are other better methods to check the validity of a run-directory.
Audit() method is now a base Job class method, as most of the variations between them were removed.

* ''' changed to """

* Adjust file filtering

Adjust file filtering to ignore 'zero-files' directories from the get-go.