Added success confirmation for child jobs in arrays (#81)
* Added success confirmation for child jobs in arrays

QCJob and FastQCJob now generate '.completed' completed files; one for
each job in the array that completes successfully. After the primary job
completes, the number of .completed files is counted and an Error is
raised if the number is less than expected.

* Updated existing tests

* Updated existing unittests

* flake8

* Added unittest for .completed file generation in QCJob

* Added unittest for .completed files for FastQCJob

* Updated unittest documentation

Updated unittest documentation to reflect what's described in Github
Workflows.

* FastQCJob now logs which array sub-jobs failed

If one or more sub-jobs in an array fail, FastQCJob will generate a file
in the logs directory that gives the indexes for all failed sub-jobs and
the job_id of the primary job itself.

* merge, flake8

* QCJobs updated to generate failed_indexes log file

QCJobs updated to generate a log file in the logs directory, containing
the job_id and index of all failed sub-jobs. This is similar to what
FastQCJob now does as well.

* flake8