Minor updates
NuQCJob's nprocs parameter removed, since it was not being used.
The standard for supplying .mmi files to NuQC is established as a path
to a directory with two subdirectories 'metatranscriptomic' and
'metagenomic'. Each directory will contain the files or symlinks to the
files needed to support the specified assay type. The creator of the
NuQCJob object will be responsible for supplying the correct path.
Hence, updates are largely documentation only.